requirements.txt
aiogram~=2.25.1
segno~=1.5.2
requests~=2.31.0
Flask~=2.2.5
pip~=23.1.2
colorama~=0.4.6
pytz~=2023.3
idna~=3.4
multidict~=6.0.4
attrs~=23.1.0
MarkupSafe~=2.1.3
Werkzeug~=2.3.6
click~=8.1.6
Jinja2~=3.1.2
itsdangerous~=2.1.2
numpy~=1.25.2
setuptools~=65.5.0
cffi~=1.15.1
six~=1.16.0
urllib3~=1.26.15
certifi~=2022.12.7
pandas~=2.0.3
openpyxl~=3.1.2
python-dotenv~=1.0.0
Pillow~=10.0.1
psycopg2-binary